Convert type_utils.py to use arraylias (#316)
