+++ test/dynamics/models/test_operator_collections.py
@@ -30,8 +30,6 @@ from qiskit_dynamics.models.operator_collections import (
     VectorizedLindbladCollection,
     ScipySparseVectorizedLindbladCollection,
 )
-from qiskit_dynamics.array import Array
-from qiskit_dynamics.type_utils import to_array
 from ..common import test_array_backends, QiskitDynamicsTestCase
 
 
@@ -171,12 +169,12 @@ class TestScipySparseOperatorCollection(QiskitDynamicsTestCase):
 
         # 2d case
         value = collection(np.array([1.0, 2.0]), np.ones((2, 2)))
-        self.assertTrue(isinstance(value, (np.ndarray, Array)))
+        self.assertTrue(isinstance(value, np.ndarray))
         self.assertAllClose(value, 3.0 * np.ones((2, 2)))
 
         # 1d case
         value = collection(np.array([1.0, 2.0]), np.array([1.0, 1.0]))
-        self.assertTrue(isinstance(value, (np.ndarray, Array)))
+        self.assertTrue(isinstance(value, np.ndarray))
         self.assertAllClose(value, np.array([3.0, 3.0]))
 
     def test_consistency_with_dense_pseudorandom(self):
@@ -207,17 +205,17 @@ class TestScipySparseOperatorCollection(QiskitDynamicsTestCase):
         for _ in range(4):
             op = r(3, 3)
             ham_ops.append(Operator(op))
-            ham_ops_alt.append(Array(op))
+            ham_ops_alt.append(op)
         sigVals = r(4)
         static_operator_numpy_array = r(3, 3)
         sparse_collection_operator_list = ScipySparseOperatorCollection(
             operators=ham_ops, static_operator=Operator(static_operator_numpy_array)
         )
         sparse_collection_array_list = ScipySparseOperatorCollection(
-            operators=ham_ops_alt, static_operator=to_array(static_operator_numpy_array)
+            operators=ham_ops_alt, static_operator=static_operator_numpy_array
         )
         sparse_collection_pure_array = ScipySparseOperatorCollection(
-            operators=to_array(ham_ops), static_operator=to_array(static_operator_numpy_array)
+            operators=ham_ops, static_operator=static_operator_numpy_array
         )
         a = sparse_collection_operator_list(sigVals)
         b = sparse_collection_array_list(sigVals)
